Move pre-print warmup pause to START_PRINT.
Prior to this change, the startup gcode that paused the print until
the target extruder temperature was reached was part of the startup
gcode in the slicer. This change moves that logic into the
START_PRINT block. Eventually I'd probably like to just do
all the necessary startup logic here and have the slicer just
emit 'PRINT_WARMUP [...]' followed by 'START_PRINT [...]'.