Rework homing to use the pressure switch as zero.
Prior to these changes, the configuration (based on the original
Sovol config) used a custom copy of the `probe` module to calibrate
the Z offset relative to a pressure switch at the corner of the bed
of the printer. The actual zero coordinate of the printer was then
set relative to the inductive probe trigger point, which is notoriously
variable.

This change reworks the configuration to use the pressure switch as
the Z endstop, and Z homing is now always done at the pressure switch.
The inductive probe is used to level the gantry and model the bed mesh,
but these coordinates are now always relative to the pressure switch
position.

This has the advantage of removing the need for a custom module for
probing and seems to produce consistent results. One caveat I have
found is that (predictably) homing needs to happen at a consistent
temperature, otherwise the Z position will vary due to thermal
expansion.